Updated Docstrings, Comments, and Type Hints
Updated and added many docstrings, comments, and type hints.
Removed unnecessary imports.

Note to self:
Still need to update readme
Also, create way to know which seeds are used for which hand numbers for reproducability. (Or better, only generate one set of decks and somehow keep track which hand sizes it was used for)