Added hand_size to name of loaded decks with seed
Added hand_size to name of loaded decks with seed so that results can be reproduced. Still may want to rework in the future so that separate deck arrays are not generated for each hand_size.

Made minor and temporary changes to main.py so I can clear all from terminal.